A few more diversions
Triangular numbers was something of a surprise.

I have a feeling there's something to the observation these can be expressed as (n(n + 1))/2.
Though, I'm not sure what this is.

After a little too much thinking, a simple brute force out of interest for how difficult, and it's very not difficult.