euler 27, quadratic primes
this was nice
some common sense suggested there was a way to cut down the search space
and, with the observation that n starts at 0, deductions for a and b became clear

the implementation is fairly inefficient, as i take an upper bound on primes, do repeated bound checks which could be elided, and then perform trial division to tests for primality.
though, perhaps there's a more efficient approach to this regardless.

troubles encountered were overlooking the base case in prime by trial
after a few failed submissions the overlooked 0 / 1 cases where fixed
of course, the test cases from the prompt passed fine without the fix…