euler 45, triangular, pentagonal, and hexagonal
Another very, very, quite, simple approach.

Set some upper bound and generate sets of the relevant numbers for indicies below the bound, and then take the intersection.

Something a little more interseting would be to work through the hexagonals (as these grow faster) and then test for triangular and hexagonal.
Whether there's a good test I'm not sure, and as I didn't make the effort for the previous problem…